refactor(shared): consolidate MCP utility functions into shared lib
isKnowledgeMcp, splitMcpByType, and the URL construction logic for
mcpToPlatformUrl/promptToPlatformUrl are now in libs/shared. The factory
functions (createMcpPlatformUrl, createPromptPlatformUrl) keep env access
out of shared; each app's utility file binds VITE_PLATFORM_URL and
re-exports under the original names so no call sites change.
