+++ packages/client/src/utility/mcp-utils.ts
@@ -1,31 +1,6 @@
-import type { MCP, Prompt } from "@semoss/shared";
+import { createMcpPlatformUrl, createPromptPlatformUrl } from "@semoss/shared";
 
-const PLATFORM_URL = import.meta.env.VITE_PLATFORM_URL
-	? import.meta.env.VITE_PLATFORM_URL
-	: "";
+const PLATFORM_URL = import.meta.env.VITE_PLATFORM_URL ?? "";
 
-/**
- * Convert the MCP into a platform url
- */
-export const mcpToPlatformUrl = (
-	mcp:
-		| Pick<MCP, "type" | "id">
-		| {
-				engine_id: string;
-				engine_type: string;
-		  },
-): string => {
-	const id = "id" in mcp ? mcp.id : mcp.engine_id;
-	const type = "type" in mcp ? mcp.type : mcp.engine_type;
-	if (type === "PROJECT") {
-		return `${PLATFORM_URL}/#/app/${id}/mcp-usage`;
-	}
-	return `${PLATFORM_URL}/#/engine/${type.toLowerCase()}/${id}/mcp-usage`;
-};
-
-/**
- * Convert the Prompt into a platform url
- */
-export const promptToPlatformUrl = (prompt: Pick<Prompt, "id">): string => {
-	return `${PLATFORM_URL}/#/prompt/${prompt.id}`;
-};
+export const mcpToPlatformUrl = createMcpPlatformUrl(PLATFORM_URL);
+export const promptToPlatformUrl = createPromptPlatformUrl(PLATFORM_URL);
